Keep separating public and private API
See: b08c787fe
